Moved timeseries to use fewer models and to be less exposed through aggregation.
    - Removed unnecessary models
    - Now in the boundary of data, only dicts are exchanged
